+++ internal/domain/Structs/Operation.go
@@ -1,10 +1,10 @@
 package domainstructs
 
 type Operation struct {
-	FirstOperand  float64 `json:",omitempty"`
-	SecondOperand float64 `json:",omitempty"`
+	FirstOperand  int64 `json:",omitempty"`
+	SecondOperand int64 `json:",omitempty"`
 }
 
 type Result struct {
-	Is float64
+	Is int64
 }
